Overhauled database retrieval and deletion methods and implemented deletion functionality on general rolls

+++ app/src/main/java/com/example/lfelne/DBHandler.java
@@ -5,7 +5,6 @@ import android.content.Context;
 import android.database.Cursor;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
-import android.util.Pair;
 
 import java.util.ArrayList;
 
@@ -14,7 +13,6 @@ public class DBHandler extends SQLiteOpenHelper {
     private static final String DB_NAME = "lfelnedb";
     private static final int DB_VERSION = 1;
     private static final String TABLE_NAME = "rolls";
-    private static final String ID_COL = "id";
     private static final String NAME_COL = "roll_name";
     private static final String TYPE_COL = "roll_type";
     private static final String MAX_EXPOSURES_COL = "max_exposures";
@@ -30,7 +28,6 @@ public class DBHandler extends SQLiteOpenHelper {
     @Override
     public void onCreate(SQLiteDatabase db) {
         String query = "CREATE TABLE " + TABLE_NAME + " ("
-                + ID_COL + " INTEGER PRIMARY KEY AUTOINCREMENT, "
                 + NAME_COL + " TEXT,"
                 + TYPE_COL + " TEXT,"
                 + MAX_EXPOSURES_COL + " TEXT,"
@@ -49,17 +46,16 @@ public class DBHandler extends SQLiteOpenHelper {
         values.put(TYPE_COL, roll.getROLL_TYPE());
         values.put(MAX_EXPOSURES_COL, roll.getMAX_EXPOSURES());
         values.put(EXPOSURES_COL, roll.getExposures());
-        values.put(START_DATE_COL, roll.getSTART_DATE().getTime());
-        values.put(END_DATE_COL, roll.getEnd_date().getTime());
+        values.put(START_DATE_COL, roll.getSTART_DATE());
+        values.put(END_DATE_COL, roll.getEnd_date());
 
         db.insert(TABLE_NAME, null, values);
         db.close();
     }
 
-    public Pair<ArrayList<Roll>, Integer> collectRolls() {
-        ArrayList<Roll> returnList = new ArrayList<Roll>();
+    public ArrayList<Roll> collectRolls() {
+        ArrayList<Roll> returnList = new ArrayList<>();
         String selectQuery = "SELECT  * FROM " + TABLE_NAME;
-        int index;
 
         SQLiteDatabase db = this.getReadableDatabase();
         Cursor cursor = db.rawQuery(selectQuery, null);
@@ -79,30 +75,12 @@ public class DBHandler extends SQLiteOpenHelper {
         }
         cursor.close();
         db.close();
-
-        index = returnList.size() - 1;
-        return new Pair<>(returnList, index);
+        return returnList;
     }
 
-    public int getLastIndex() {
-        int returnInt = -1;
+    public Roll getRoll(long start_date) {
         SQLiteDatabase db = this.getReadableDatabase();
-        String selectQuery = "SELECT  * FROM " + TABLE_NAME;
-        Cursor cursor = db.rawQuery(selectQuery, null);
-
-        if (cursor.moveToLast())
-            returnInt = cursor.getPosition();
-        cursor.close();
-        db.close();
-
-        return returnInt;
-    }
-
-    public Roll getRoll(int index) {
-        SQLiteDatabase db = this.getReadableDatabase();
-        String selectQuery = "SELECT  * FROM " + TABLE_NAME;
-        Cursor cursor = db.rawQuery(selectQuery, null);
-        cursor.moveToPosition(index);
+        Cursor cursor = generateCursor(db, start_date);
 
         String roll_init = cursor.getString(1) + "-"
                 + cursor.getString(2) + "-"
@@ -117,9 +95,22 @@ public class DBHandler extends SQLiteOpenHelper {
         return roll;
     }
 
+    public void removeRoll(long start_date) {
+        String deleteQuery = "DELETE FROM " + TABLE_NAME + " WHERE start_date = " + start_date;
+        SQLiteDatabase db = this.getWritableDatabase();
+        db.execSQL(deleteQuery);
+    }
+
+    public Cursor generateCursor(SQLiteDatabase db, long start_date) {
+        String selectQuery = "SELECT  * FROM " + TABLE_NAME + " WHERE start_date = " + start_date;
+        return db.rawQuery(selectQuery, null);
+    }
+
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
-        onCreate(db);
+        if (newVersion > oldVersion) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
+            onCreate(db);
+        }
     }
 }
